+++ entities/entity.cpp
@@ -21,14 +21,14 @@ void Entity::wire_renderer_to_body()
     m_renderer->onUpdatePos(m_body->x(), m_body->y());
     m_renderer->onUpdateSize(m_body->width(), m_body->height());
 
-    connect(m_body, SIGNAL(positionChanged(int,int)), m_renderer, SLOT(onUpdatePos(int,int)));
-    connect(m_body, SIGNAL(sizeChanged(int,int)), m_renderer, SLOT(onUpdateSize(int,int)));
+    connect(m_body, SIGNAL(positionChanged(qreal,qreal)), m_renderer, SLOT(onUpdatePos(qreal,qreal)));
+    connect(m_body, SIGNAL(sizeChanged(qreal,qreal)), m_renderer, SLOT(onUpdateSize(qreal,qreal)));
 }
 
 void Entity::unwire_renderer_from_body()
 {
-    disconnect(m_body, SIGNAL(positionChanged(int,int)), m_renderer, SLOT(onUpdatePos(int,int)));
-    disconnect(m_body, SIGNAL(sizeChanged(int,int)), m_renderer, SLOT(onUpdateSize(int,int)));
+    disconnect(m_body, SIGNAL(positionChanged(qreal,qreal)), m_renderer, SLOT(onUpdatePos(qreal,qreal)));
+    disconnect(m_body, SIGNAL(sizeChanged(qreal,qreal)), m_renderer, SLOT(onUpdateSize(qreal,qreal)));
 }
 
 void Entity::dispose_of_renderer()
